requirements.txt
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.38.0
wikipedia==1.4.0
wolframalpha==5.1.3
